Clean up OFFSPEC notebook
