+++ src/ess/external/offspec/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+
+from . import data
+
+__all__ = ['data']
